Refactor: 웹 접근성 개선 - 버튼 및 링크에 aria-label로 접근 가능한 이름 추가 - <ul> 요소 안에 <li> 요소가 아닌 <a> 요소가 있던 문제 수정
