+++ src/app/blog/page.tsx
@@ -1,6 +1,6 @@
 //카테고리를 아예 입력 안했을 때
 
-import PostListPage from "@/components/post_list/PostListPage2";
+import PostListPage from "@/components/post_list/PostListPage";
 
 const BlogPage = async () => {
     return <PostListPage />;
