+++ src/layouts/Footer.tsx
@@ -6,10 +6,18 @@ export default function Footer() {
     return (
         <footer className="flex flex-col justify-center items-center text-center mb-16 mt-20 gap-4 print:hidden">
             <div className="flex justify-center items-center gap-4 ">
-                <Link href="mailto:swhan0428@gmail.com" target="_blank">
+                <Link
+                    href="mailto:swhan0428@gmail.com"
+                    target="_blank"
+                    aria-label="Mailto"
+                >
                     <Mail className="hover:stroke-blue-500" />
                 </Link>
-                <Link href="https://github.com/swhana" target="_blank">
+                <Link
+                    href="https://github.com/swhana"
+                    target="_blank"
+                    aria-label="Github"
+                >
                     <GithubIcon
                         className="fill-current hover:fill-blue-500 dark:fill-white"
                         width={24}
@@ -19,6 +27,7 @@ export default function Footer() {
                 <Link
                     href="https://www.linkedin.com/in/sangwon-han"
                     target="_blank"
+                    aria-label="Linkedin"
                 >
                     <Linkedin
                         width={24}
